+++ packages/shell/src/lib/components/ui/tabs-content.svelte
@@ -1,8 +1,0 @@
-<script lang="ts">
-	import { Tabs as TabsPrimitive } from "bits-ui";
-	import { cn } from "../../utils.ts";
-
-	let { class: className, ...restProps }: TabsPrimitive.ContentProps = $props();
-</script>
-
-<TabsPrimitive.Content class={cn("mt-4 outline-none", className)} {...restProps} />